Create c_program_to_find_sum_of_two_numbers.c
In this program we are using two printf and scanf functions for taking two inputs from use \n is for next line of statement.
In place of using two seperate we can use one function icluding both inputs this is equavalent to the same and works same as written. 
when using one we put %d%d for taking two inputs and scaning both in one line. for this we will discuss in future programs in same repositery.