Create printing_a_number_100_times.c
loop is a function that is used to repeat a function for many times as  here the for loop is used and there are while loop and do while loop   present in c language. here the loop will run the printf statement as many time as i entered && untill the i = 100.